Consolidate some fuzzers into one
This commit consolidates a number of fuzzers that we have in Wasmtime
into a single fuzzer. This follows the pattern of [wasm-tools's `run`
fuzzer][tools] where the first byte in the fuzz input says which fuzzer
to run and then the rest of the input is the input to the fuzzer. This
is intended to improve our time-slicing on OSS-Fuzz where less important
fuzzers don't hog time from the "important fuzzers" such as
`differential`. Some minor-ish fuzzers have been folded in to this new
fuzzer, but we can also always move things around as needed

[tools]: https://github.com/bytecodealliance/wasm-tools/blob/main/fuzz/fuzz_targets/run.rs
