+++ fuzz/Cargo.toml
@@ -58,12 +58,6 @@ test = false
 doc = false
 
 [[bin]]
-name = "api_calls"
-path = "fuzz_targets/api_calls.rs"
-test = false
-doc = false
-
-[[bin]]
 name = "differential"
 path = "fuzz_targets/differential.rs"
 test = false
@@ -72,18 +66,6 @@ doc = false
 [[bin]]
 name = "wast_tests"
 path = "fuzz_targets/wast_tests.rs"
-test = false
-doc = false
-
-[[bin]]
-name = "table_ops"
-path = "fuzz_targets/table_ops.rs"
-test = false
-doc = false
-
-[[bin]]
-name = "stacks"
-path = "fuzz_targets/stacks.rs"
 test = false
 doc = false
 
@@ -118,22 +100,8 @@ test = false
 doc = false
 
 [[bin]]
-name = "memory_accesses"
-path = "fuzz_targets/memory_accesses.rs"
-test = false
-doc = false
-bench = false
-
-[[bin]]
-name = "pulley"
-path = "fuzz_targets/pulley.rs"
-test = false
-doc = false
-bench = false
-
-[[bin]]
-name = "dominator_tree"
-path = "fuzz_targets/dominator_tree.rs"
+name = "misc"
+path = "fuzz_targets/misc.rs"
 test = false
 doc = false
 bench = false
