--- fuzz/fuzz_targets/api_calls.rs
@@ -1,8 +1,0 @@
-#![no_main]
-
-use libfuzzer_sys::fuzz_target;
-use wasmtime_fuzzing::{generators::api::ApiCalls, oracles};
-
-fuzz_target!(|api: ApiCalls| {
-    oracles::make_api_calls(api);
-});
