--- fuzz/fuzz_targets/memory_accesses.rs
@@ -1,8 +1,0 @@
-#![no_main]
-
-use libfuzzer_sys::fuzz_target;
-use wasmtime_fuzzing::{generators::MemoryAccesses, oracles::memory::check_memory_accesses};
-
-fuzz_target!(|input: MemoryAccesses| {
-    check_memory_accesses(input);
-});
