--- fuzz/fuzz_targets/table_ops.rs
@@ -1,9 +1,0 @@
-#![no_main]
-
-use libfuzzer_sys::fuzz_target;
-use wasmtime_fuzzing::generators::{table_ops::TableOps, Config};
-
-fuzz_target!(|pair: (Config, TableOps)| {
-    let (config, ops) = pair;
-    let _ = wasmtime_fuzzing::oracles::table_ops(config, ops);
-});
